Move mkosi.py to src/ directory.
This allows us to start splitting mkosi up into multiple files.
Releasing can be done using zipapp which bundles all the separate files
into a single python executable zip archive that can be uploaded to
Github.
