+++ do-a-release.sh
@@ -6,11 +6,13 @@ if [ x"$1" == x ] ; then
     exit 1
 fi
 
-sed -ie 's/version=".*",/version="'"$1"'",/' setup.py
-sed -ie "s/__version__ = '.*'/__version__ = '$1'/" mkosi
+sed -ie "s/__version__ = '.*'/__version__ = '$1'/" src/mkosi.py
 
-git add -p setup.py mkosi
+git add -p src/mkosi.py
 
 git commit -m "bump version numbers for v$1"
 
 git tag -s "v$1" -m "mkosi $1"
+
+mkdir -p build && python3 -m zipapp src -m mkosi:main -p "/usr/bin/env python3" -o build/mkosi
+echo "Add build/mkosi as a release artifact to the release on Github!"
