--- mkosi.py
@@ -1,1 +1,0 @@
-mkosi
